+++ src/com/lilithsthrone/game/dialogue/utils/UtilText.java
@@ -795,11 +795,11 @@ public class UtilText {
 		return formatAsMoney(moneyString, "b", PresetColour.CURRENCY_COPPER);
 	}
 	
-	public static String formatAsMoney(int money) {
+	public static String formatAsMoney(long money) {
 		return formatAsMoney(money, "b");
 	}
 	
-	public static String formatAsMoneyUncoloured(int money, String tag) {
+	public static String formatAsMoneyUncoloured(long money, String tag) {
 		return formatAsMoney(money, tag, null);
 	}
 
@@ -807,7 +807,7 @@ public class UtilText {
 		return formatAsMoney(money, tag, null);
 	}
 	
-	public static String formatAsMoney(int money, String tag) {
+	public static String formatAsMoney(long money, String tag) {
 		return formatAsMoney(money, tag, PresetColour.TEXT);
 	}
 	
@@ -829,7 +829,7 @@ public class UtilText {
 		return formatAsMoney(money, tag, PresetColour.TEXT);
 	}
 	
-	public static String formatAsMoney(int money, String tag, Colour amountColour) {
+	public static String formatAsMoney(long money, String tag, Colour amountColour) {
 		return formatAsMoney(Units.number(money), tag, amountColour);
 	}
 	
@@ -1608,7 +1608,7 @@ public class UtilText {
 				"Formats the passed integer argument as a String.") {
 			@Override
 			public String parse(List<GameCharacter> specialNPCs, String command, String arguments, String target, GameCharacter character) {
-				return Util.intToString(Integer.valueOf(UtilText.parse(arguments)));
+				return Util.intToString(Long.valueOf(UtilText.parse(arguments)));
 			}
 		});
