+++ src/com/lilithsthrone/game/inventory/CharacterInventory.java
@@ -70,7 +70,7 @@ public class CharacterInventory implements XMLSaving {
 	private final Map<String, List<InventorySlot>> unlockKeyMap;
 
 	protected int essenceCount;
-	protected int money;
+	protected long money;
 	
 	private Set<InventorySlot> dirtySlots;
 	
@@ -284,7 +284,7 @@ public class CharacterInventory implements XMLSaving {
 			inventory.floorInventory = CharacterInventory.loadingFromFloorBackupCheck;
 		}
 		
-		inventory.setMoney(Integer.valueOf(((Element)parentElement.getElementsByTagName("money").item(0)).getAttribute("value")));
+		inventory.setMoney(Long.valueOf(((Element)parentElement.getElementsByTagName("money").item(0)).getAttribute("value")));
 		
 		if(parentElement.getElementsByTagName("essences").item(0)!=null) { // Old version support.
 			inventory.setEssenceCount(Integer.valueOf(((Element)parentElement.getElementsByTagName("essences").item(0)).getAttribute("value")));
@@ -463,25 +463,25 @@ public class CharacterInventory implements XMLSaving {
 				&& clothingCurrentlyEquipped.isEmpty();
 	}
 
-	public int getMoney() {
+	public long getMoney() {
 		return money;
 	}
 
 	/**
 	 * Does not allow money to fall below 0.
 	 */
-	public void setMoney(int newValue) {
+	public void setMoney(long newValue) {
 		money = Math.max(0, newValue);
 	}
 	
 	/**
 	 * Does not allow money to fall below 0.
 	 */
-	public void incrementMoney(int increment) {
+	public void incrementMoney(long increment) {
 		try {
 			setMoney(Math.addExact(money, increment));
 		} catch (ArithmeticException ex) {
-			setMoney(Integer.MAX_VALUE);
+			setMoney(Long.MAX_VALUE);
 		}
 	}
